Crate axprocess

This module provides the process management API for the operating system.

Modules

flags

clone 任务时指定的参数。

futex

实现与futex相关的系统调用

link

模拟的链接、挂载模块 fat32本身不支持符号链接和硬链接，两个指向相同文件的目录条目将会被chkdsk报告为交叉链接并修复

signal

负责处理进程中与信号相关的内容

Structs

Process

The process control block

Statics

PID2PC

Map from process id to arc pointer of process

TID2TASK

Map from task id to arc pointer of task

Functions

current_process

return the Arc<Process> of the current process

current_task

current running task

exit_current_task

退出当前任务

get_task_ref

Get the task reference by tid

handle_page_fault

To deal with the page fault

init_kernel_process

初始化内核调度进程

load_app

返回应用程序入口，用户栈底，用户堆底

set_child_tid

设置当前任务的clear_child_tid

sleep_now_task

以进程作为中转调用task的sleep

time_stat_from_kernel_to_user

当从内核态到用户态时，统计对应进程的时间信息

time_stat_from_user_to_kernel

当从用户态到内核态时，统计对应进程的时间信息

time_stat_output

统计时间输出 (用户态秒，用户态微秒，内核态秒，内核态微秒)

wait_pid⚠

在当前进程找对应的子进程，并等待子进程结束 若找到了则返回对应的pid 否则返回一个状态

yield_now_task

以进程作为中转调用task的yield