Module axprocess::link

模拟的链接、挂载模块 fat32本身不支持符号链接和硬链接，两个指向相同文件的目录条目将会被chkdsk报告为交叉链接并修复

Structs

FilePath

A struct to represent a file path, which will be canonicalized

Constants

AT_FDCWD

The file descriptor used to specify the current working directory of a process

Statics

LINK_COUNT_MAP

实际文件(而不是用户文件)到链接数的映射

LINK_PATH_MAP

用户看到的文件到实际文件的映射

Functions

create_link

创建一个链接

deal_with_path

To deal with the path and return the canonicalized path

get_link_count

获取文件的链接数

get_str_len⚠

Safety

raw_ptr_to_ref_str⚠

Safety

real_path

将用户提供的路径转换成实际的路径

remove_link

删除一个链接